Clear compiler warnings in library, helper and example code
Enable -Wall -Wextra (GCC/Clang, via the ENABLE_WARNINGS option) and fix
everything it reports outside the HDF5 output code and the tests.

krylovObservables.cpp compared the int parameter 'len' against the size_t
member 'dim' and iterated over the size_t 'numSamples' with an int, both
of which relied on implicit signed/unsigned conversion.

createNumberOperatorObservables iterated with an unsigned int over the
signed numberModes.

simpleExample declared 'tol' and 'm' next to a comment explaining them,
but then called the five-argument krylovTimeEvolver constructor, which
takes neither. The values happened to match the defaults that the short
constructor delegates (1e-6 and 40), so the numerics were correct, but
editing them in the example had no effect whatsoever. Drop the variables
and document what the short constructor actually does. Also drop the
unused nbObservables.

The CSV writing loop in simpleExample computed retNumSamples() - 1 in
size_t arithmetic, which underflows to SIZE_MAX for an observable with no
samples. Read the count once and guard the empty case.

The warnings left in krylovHelper.cpp (deprecated H5::Attribute copy
assignment) and in test/ are deliberately untouched; both areas are due
for rework.

+++ core/krylovObservables.cpp
@@ -60,7 +60,7 @@ krylovBasicObservable::krylovBasicObservable(const std::string& name, std::vecto
 {
     expectationValues = new double[numSamples];
 
-    for (int i = 0; i != numSamples; i++) {
+    for (size_t i = 0; i != numSamples; i++) {
         expectationValues[i] = values[i];
     }
 }
@@ -123,7 +123,7 @@ std::complex<double> krylovMatrixObservable::expectation(std::complex<double>* v
         std::cerr << "Too many samples." << std::endl;
         exit(1);
     }
-	if (len != dim)
+	if (static_cast<size_t>(len) != dim)
 	{
 		std::cerr << "Incompatible dimensions" << std::endl;
 		exit(1);
@@ -177,7 +177,7 @@ std::complex<double> krylovSpMatrixObservable::expectation(std::complex<double>*
         std::cerr << "Too many samples." << std::endl;
         exit(1);
     }
-	if (len != dim)
+	if (static_cast<size_t>(len) != dim)
 	{
 		std::cerr << "Incompatible dimensions" << std::endl;
 		exit(1);
@@ -220,7 +220,7 @@ std::complex<double> krylovVectorObservable::expectation(std::complex<double>* v
         std::cerr << "Too many samples." << std::endl;
         exit(1);
     }
-	if (len != dim)
+	if (static_cast<size_t>(len) != dim)
 	{
 		std::cerr << "Incompatible dimensions" << std::endl;
 		exit(1);
@@ -244,7 +244,7 @@ krylovOutputObservable::krylovOutputObservable(const std::string& name, std::vec
 {
     this->numSamples = values.size();
     initializeResultArray(numSamples);
-    for (int i = 0; i != numSamples; i++) {
+    for (size_t i = 0; i != numSamples; i++) {
         expectationValues[i] = values[i];
     }
 }
